function that creates a file

+++ 0x14-file_io/1-create_file.c
@@ -11,7 +11,7 @@
 
 int create_file(const char *filename, char *text_content)
 {
-	ssize_t file;
+	int file;
 	int count = 0;
 
 	if (!filename)
